+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "function-trigger-deployer-rs",
     "function-deployer-rs",
     "function-snapstart-deployer-rs",
+    "common_lib"
 ]
 
 [profile.release]
