+++ common_lib/Cargo.toml
@@ -5,3 +5,27 @@ edition = "2021"
 
 [dependencies]
 serde = "1.0.179"
+
+
+
+#serde = "1.0.179"
+#serde_json = "1.0.104"
+#tracing-subscriber = { version = "0.3.17", features = ["json"] }
+
+#testcontainers = "0.15.0"
+async-trait = "0.1.74"
+aws-sdk-lambda = "0.34.0"
+#aws-sdk-dynamodb = "0.34.0"
+#aws-config = "0.56.1"
+tracing = { version = "0.1.40", features = ["log"] }
+#aws-sdk-sqs = "0.34.0"
+#rand = "0.8.5"
+aws-sdk-cloudwatchlogs = "0.34.0"
+
+# TODO: stop using error from lambda_runtime
+lambda_runtime = "0.8.1"
+aws-config = "0.56.1"
+
+# TODO check if we can get rid of it
+tokio = { version = "1", features = ["full"] }
+rand = "0.8.5"
